requirements.txt
# 核心依赖
openai>=1.0.0
requests>=2.31.0
psutil>=5.9.0

# 图像和多媒体
pillow>=10.0.0
pygame>=2.5.0
opencv-python>=4.8.0

# 自动化工具
pyautogui>=0.9.54
pynput>=1.7.6

# 系统相关（Windows）
pywin32>=306

# 网络和API
websocket-client>=1.6.0
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
sqlite3
json

# 多媒体
pydub>=0.25.1

# GUI界面
tkinter
customtkinter>=5.2.0
pygame-gui>=0.6.0

# 其他工具
schedule>=1.2.0
python-dotenv>=1.0.0
logging
asyncio
threading